rrdcached: fix init script "stop": wait until termination
The initscript just sent the kill signal without waiting
of the daemon to actually stop. That causes problems
in several situation, e.g. a restart might fail or
a scripts using 'omd stop' might assume that the daemon
is stopped while it isn't.

Currently there is a timeout of 20 seconds. If the
rrdcached does not terminate within this timeout,
the init script will output an error message and
exit with 1.