pnp4nagios: remove symlink var/log/pnp4nagios
Reason: the older versions do not support symlinks to
directories in skel. They crash on an omd update. So
that if we introduce this link *now*, users will not
be able to downdate from 0.46 to 0.44. So we rather wait
for 0.48 or 0.50 to introduce this symlink when older
versions then support such symlinks.