omd: add build-cached target
add new build-cached target which drastically increases build speed. You have to
create an (empty) folder anywhere accessible by the build machine. Then add
a new file '.config' to the repository and add 'BUILD_CACHE=<folder>' to that
.config file. Then the build result will be cached and reused on the next build
until something is changed in that package.
You have to clear the cache manually when the version or build root changes.

Signed-off-by: Sven Nierlein <Sven.Nierlein@consol.de>