avoid merge conflicts
if a skel file has the sitename and a config value changed from a hook then it
will break on updates because the value is changed from a hook but omd trys to
patch the sitename on move/copys.
adding ###CONFIG_...### replacements to skel will make the patches work and will
also make "omd diff" be empty on others than the first site.