update prometheus to 2.37.5
The build process has changed a bit. The static web content is not
compiled into a single go file now. It stays on the filesystem and
is later ingested during compile time. The embed.go just contains
references to the files on the filesystem now.

That is why the tar ball was much smaller when we tried the naive
upgrade. Because we deleted that static files.